dan/pyproject.toml
[tool.poetry]
name = "3dprint.dan"
version = "0.1.0"
description = ""
authors = ["Daniel Winkler <dan@danwink.com>"]

[tool.poetry.dependencies]
python = "~3.9"
plumbum = "~1.7.0"
numpy = "~1.20.2"
numba = "*"
solidpython = "*" #{git = "https://github.com/danwinkler/SolidPython"}
PyMCubes = "~0.1.2"
scipy = "^1.6.2"
# cython-csg = "^0.5.8"
numpy-stl = "^2.16.0"
tqdm = "^4.60.0"
pyclipper = "^1.3.0"
shapely = "^1.7.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.12.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
